pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datapact"
version = "2.0.0"
description = "DataPact: An enterprise-grade, programmatic data validation accelerator for Databricks."
readme = "README.md"
requires-python = ">=3.13.5"
authors = [
  { name = "Skyler Myers", email = "myers_skyler@icloud.com" },
]
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.13",
    "Development Status :: 5 - Production/Stable",
]
dependencies = [
    "databricks-sdk>=0.62.0",
    "databricks-sql-connector",
    "pyarrow",
    "pyyaml",
    "loguru",
    "pydantic",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pre-commit"
]

[project.urls]
"Homepage" = "https://www.entrada.ai/datapact"
"Documentation" = "https://docs.entrada.ai/datapact"

[project.scripts]
datapact = "datapact.main:main"

[tool.setuptools.packages.find]
include = ["datapact*"]
